Move onewireeentity.py to onewireeentities.py
