+++ homeassistant/components/onewire/onewireeentities.py
